fix: restructure codebase to be lean :octocat:
