+++ src/storage/base.ts
@@ -1,0 +1,4 @@
+export abstract class Storage<T> {
+  abstract put(record: T): void;
+  abstract get(): T[];
+}
