+++ src/worker.ts
@@ -1,0 +1,3 @@
+export abstract class Worker<T, U> {
+  abstract work(records: T[]): Promise<U[]>;
+}
